up-to-date

+++ src/um/QuotaDatastore.cc
@@ -1,5 +1,5 @@
 /* -------------------------------------------------------------------------- */
-/* Copyright 2002-2016, OpenNebula Project, OpenNebula Systems                */
+/* Copyright 2002-2018, OpenNebula Project, OpenNebula Systems                */
 /*                                                                            */
 /* Licensed under the Apache License, Version 2.0 (the "License"); you may    */
 /* not use this file except in compliance with the License. You may obtain    */
@@ -32,7 +32,7 @@ bool QuotaDatastore::check(Template * tmpl, Quotas& default_quotas, string& erro
     map<string, float> ds_request;
 
     string ds_id;
-    int    size;
+    float  size;
     int    images, images_req = 1;
 
     tmpl->get("DATASTORE", ds_id);
@@ -68,7 +68,7 @@ void QuotaDatastore::del(Template * tmpl)
     map<string, float> ds_request;
 
     string ds_id;
-    int    size;
+    float  size;
     int    images, images_req = 1;
 
     tmpl->get("DATASTORE", ds_id);
